Implemented character experience (no way to earn it yet). Updated TODO list.

+++ player.h
@@ -31,7 +31,7 @@ public:
   const char *Name() { return name.c_str(); }
   void AddChar(Object *);
   int Accomplish(unsigned long);
-  unsigned long Exp() { return exp; }
+  int Exp() { return exp; }
 
 private:
   map<string, Object *> body;
@@ -39,7 +39,7 @@ private:
   Object *room, *creator;
   unsigned long flags;
   set<unsigned long> completed;
-  unsigned long exp;
+  int exp;
 
   friend Player *player_login(string name, string pass);
   friend Player *get_player(string name);
